// src/app/models/scheduled-workout/scheduled-workout-record.model.ts
/** ScheduledWorkoutRecord frontend model. */
export interface ScheduledWorkoutRecord {
  userWorkoutId?: number;
  user_workout_id?: number;
  programWorkoutId?: number;
  program_workout_id?: number;
  workoutId?: number;
  workout_id?: number;
  programId?: number;
  program_id?: number;
  scheduledAt?: string | null;
  scheduled_at?: string | null;
  completed?: boolean | null;
  workoutName?: string | null;
  workout_name?: string | null;

}